Update gui.py
spent more time working on improving the look of the gui. mostly there now. I couldnt get the prompt in the response area to be in a grey box with rounded edges so i comprimised with bold text for now. I would still like to fix this but i'd rather start hooking the model up now