added database.py and modified gui.py
added a database.py with functions to manage a database for storing conversations. added a side bar to the gui to show all saved conversations. started working on a notification that the model is currently thinking so the user knows the model is doing something. previously there was often delays where the gui would do nothing until the model had generated a full response